+++ packages/e2b/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @computesdk/e2b
 
+## 1.7.55
+
+### Patch Changes
+
+- Updated dependencies [0732fed]
+  - computesdk@4.1.6
+  - @computesdk/provider@2.1.7
+
 ## 1.7.54
 
 ### Patch Changes